Rename SampleFileSizeWarning to FileSizeWarning
