+++ client/src/js/samples/components/FileSizeWarning.js
